backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic[email]
python-dotenv
passlib[bcrypt]
bcrypt<4
python-jose[cryptography]
python-multipart
aiofiles
alembic
pytest
httpx<0.25
